Add support for `np.bool_` scalars.
Numpy booleans (`np.bool_`) are not the same type as Python's `bool`. In
particular, `isinstance(np.bool_(True), bool) == False` which was not caught in
our encoding process (and we did not have an associated unit test to check
that).
This change adds such support and it also moves boolean encoding to a later
stage, after floats and ints which are the most common types used in ML. This
should make encoding a bit faster for these common types but slightly slower
for rarer ones.

PiperOrigin-RevId: 387843384
Change-Id: Ib6363d8ced59abb466bb7c88e0e5354ec5595fd9