Add C++ support for `float_values_buffer`+ add cross-language test.
This CL adds support for C++ to read and write from/to `float_values_buffer`.
This was already possible in the Python version, but we missed it in C++. This
allows reading and writing big float arrays to be much faster.

Besides this missing functionality, this CL also adds a rather complex unit
test that writes payloads in Python (to `float_values_buffer`) and then reads
them back from C++. This should prevent bugs like this in the future. We can
optionally add another test to write in C++ and read from Python, but because
this CL is already big, I have decided to keep it just one-way for now.

PiperOrigin-RevId: 395477163
Change-Id: I7799bc9809b97431a812bb0d1eb673d85f9e2dc2