Allow reading arbitrary protobuf messages from Python.
While the C++ backend already supported reading and writing any protocol buffer
message that's statically compiled, the Python version supported writing
arbitrary python messages but only supported reading `Data` messages. The
problem is that we can't compile the C++ extension to accept (and return) all
possible protobuf messages as that's a compile-time template parameter, and on
top of that we can't return `proto2::Message` object as that's an abstract
class.

This CL adds support for that by letting clients (e.g. Python) instantiate the
object and accepting `proto2::Message` in the interface as an opaque argument.
This changes a little how users use the API in Python, but it allows using any
proto message, including even dynamically generated ones.

PiperOrigin-RevId: 399236193
Change-Id: I6d57fecea9bada51a3b6c54a41e4057fce2ea159