Separate Dockerfile for all Python versions in a separate file.
Our internal integration tests frequently time out when running the previous
`Dockerfile` build because some operations there such as bdist for grpcio are
really slow. We can keep that version in a separate `Dockerfile.allpyversions`
file, but we can keep a leaner one with Py 3.6 (which is the most annoying to
support anyway).

PiperOrigin-RevId: 406854255
Change-Id: Ibec18af6467b40f7af4e1592cc570e6a5472b4d5