Fix crashing issue with scheduler that skips first step.
A scheduler that skips the first timestep breaks an invariant in EnvLogger (and
in RL in general). Episodes may have undefined endpoints (e.g. in non-episodic
settings or when a job is preempted or when it crashes), but they cannot have
an undefined beginning. Schedulers should never skip the first timestep given
that we use them as the reference point when iterating over episodes.

This change does a few things:

* Modifies the signature of both `RiegeliShardReader::AddStep()` and
    `RiegeliDatasetWriter::AddStep()` from returning `void` to returning
    `bool`. This allows the process to not die in C++ so that it can be caught
    in Python and then handled via an exception.
* Adds a unit test that ensures the exception is raised if a scheduler that
    skips the first timestep is used with the Riegeli backend.
* Adds a unit test that demonstrates how to store only episodic metadata by
    having the scheduler return `True` only in the first timestep.
* Adds a note to `backend_writer.py` stating that `scheduler` should never skip
    the first timestep of the episode.

PiperOrigin-RevId: 426901881
Change-Id: I625c9ec31ee8359fe5cfdc06024808f3b0ce9c75