Add `RiegeliDatasetReader::Clone()` for "cheap" multithreading.
This change adds a new method to `RiegeliDatasetReader` to allow cheaply
copying it to a new object which can be used in a new thread. Its `pybind11`
class was also modified, allowing Python consumers to also benefit from this
functionality.

This closes #2 and it provides an example of multiple read threads in
`environment_logger_test.py` (`test_parallel_readers()`).

PiperOrigin-RevId: 429297367
Change-Id: I4d6244983967c11424a81a6e60ea96498f5d7d0f