Build for Python `3.10` by default in `Dockerfile`.
This change modifies `Dockerfile` and `WORKSPACE.bazel` to build for Python
`3.10` instead of `3.8`, besides upgrading a lot of of the underlying
dependencies. All the `BUILD.bazel` files have also been updated to reflect
this. The `Dockerfile` now also takes a build `ARG` which allows us to remove
`Dockerfile.allpyversions`.

The `pip` dependencies are now managed in a different way and are installed in
the system by `pip` instead of via `bazel`'s `pip_install()` (deprecated) /
`pip_parse()`. This makes our code a lot simpler for now, but we may revisit
this in the future.

Huge thanks to Yicheng Luo <ethanluoyc@gmail.com> (ethanluoyc@) for debugging a
lot of the issues here, especially with `bazel` and `pip`. Please see https://github.com/deepmind/envlogger/pull/5 and https://github.com/deepmind/envlogger/issues/4.

This fixes issue #4.

The new PyPi release is available at https://pypi.org/project/envlogger/1.1/.

PiperOrigin-RevId: 546115546
Change-Id: I49bb7a35b029e6b925ec4fc2c029a93972a0fe5c