#dataview Support chaining DataView accessors.
Previously, `data_view["key"]` or `data_view[2]` returns a pointer to the
underlying `envlogger::Data` object.  This prevents chaining the accessors,
like`data_view["key"][2]`.  We refactor the interface such that the
`operator[]` returns a `DataView` instead of a pointer to the underlying
`envlogger::Data` object.

There are also other minor changes due to the above interface change:
- The `DataView` will crash when initialized with nullptr.  We can take a
  reference in this case, but we want to minimize the interface change in this
  CL.  We can change the interface later.
- We move one-liner definitions to header files.
- `*data_view` returns a const reference to the underlying `envlogger::Data`
  object.
- `data_view->` returns a pointer to the underlying `envlogger::Data` object.
- Remove `Type` method becase we can do `data_view->value_case()`.

PiperOrigin-RevId: 576865175
Change-Id: I0a6eaee5a18cbf527ad62c83679bdc535aaeff68