Fix codec test broken in NumPy `2.3.0`.
This test case checks that encoding unsupported payloads (e.g. modules, sets,
functions etc) raises an exception. The testing strategy is still sound, but it
broke with the update to NumPy `2.3.0` because the exception type changed from
`TypeError` to `ValueError` for modules specifically.
Instead of trying to switch to `ValueError` for modules, which introduces a bit
of complexity to the test to detect the numpy version, I decided to remove the
check to serialize modules since we never really encountered that in our
use-cases really.

PiperOrigin-RevId: 788548247
Change-Id: I0befcbf10632f1f2553b5edf731c8559f2edc66e